Fix overflow errors with large opacity grids in shared memory.
This is embarrassing.

+++ transit/src/opacity.c
@@ -564,7 +564,7 @@ attachopacity(struct transit *tr){ /* transit struct                        */
   op->Nlayer = oh->Nlayer;
   op->Nmol = oh->Nmol;
 
-  int main_shm_size  /* Size of the main shared memory, starting with: grid */
+  size_t main_shm_size  /* Size of the main segment, starting with: grid    */
     = sizeof(PREC_RES) * op->Nmol * op->Ntemp * op->Nlayer * op->Nwave
     + sizeof(int) * op->Nmol          /* op->molID  */
     + sizeof(PREC_RES) * op->Ntemp    /* op->temp   */
